requirements.txt
django-cms<3.5
Django<1.11

Pillow
cmsplugin-filer
django-filer
djangocms_column
djangocms_link
djangocms_style
django_select2
djangocms_snippet
djangocms_googlemap
djangocms_video
djangocms-text-ckeditor
cmsplugin-plain-text

django-cors-headers
django-libsass
django-analytical
django-paintstore
django_htmlmin
django_compressor
csscompressor

